Fix/MX-95 GDPR right to be forgotten (#502)
* Return 404 if member is missing

* Only try to delete IdentityEntity if it exists

+++ member-service/src/main/java/com/hedvig/memberservice/identity/IdentityEventListener.kt
@@ -53,7 +53,9 @@ class IdentityEventListener(
 
     @EventHandler
     fun on(e: MemberDeletedEvent) {
-        repository.deleteById(e.memberId)
+        if (repository.findById(e.memberId).isPresent) {
+            repository.deleteById(e.memberId)
+        }
     }
 }
